requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
bootstrap4==0.1.0
Brotli==1.1.0
cairocffi==1.7.0
CairoSVG==2.7.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
config-oc==1.0.4
cssselect2==0.7.0
defusedxml==0.7.1
Django==5.2.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-ckeditor5==0.0.4
django-cors-headers==4.3.1
django-dirtyfields==1.9.3
django-environ==0.11.2
django-js-asset==2.2.0
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-taggit==5.0.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
email-smtp==1.0.0
fonttools==4.53.0
frozenlist==1.4.1
html5lib==1.1
idna==3.7
jobject==1.0.2
jsonb==1.0.0
kombu==5.3.7
msgpack==1.0.8
multidict==6.0.5
Pillow==10.1.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycairo==1.26.0
pycparser==2.22
pydyf==0.10.0
PyGObject==3.48.2
PyJWT==2.8.0
PyMySQL==1.1.1
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2024.1
redis==5.0.4
reportlab==4.2.0
requests==2.31.0
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.3
tinycss2==1.3.0
tools-oc==1.2.4
twilio==9.1.0
tzdata==2024.1
undefined-oc==1.0.0
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
yarl==1.9.4
zopfli==0.2.3
